Consolidate KtxUtils into Extensions; new JAR
